Create a new "Stats" page with a selector for different time ranges
This includes most of the stats I could think of, as well as a
lot of stats people have asked for from Slack.

As people collect more and more numbers, fun stats like this
are going to get harder to calculate independently, so hopefully
this also acts as a framework for other interesting stats as
people come up with them.